10.3 工单模块修复，扫码功能修改，生成二维码域名修改，巡检任务修改

+++ app/src/main/java/com/hwc/dwcj/entity/second/MessageInfo.java
@@ -8,7 +8,7 @@ package com.hwc.dwcj.entity.second;
 public class MessageInfo {
 
     private String id;
-    private int msgId;
+    private long msgId;
     private String receiverId;
     private int msgState;
     private int isDel;
@@ -37,11 +37,11 @@ public class MessageInfo {
         this.id = id;
     }
 
-    public int getMsgId() {
+    public long getMsgId() {
         return msgId;
     }
 
-    public void setMsgId(int msgId) {
+    public void setMsgId(long msgId) {
         this.msgId = msgId;
     }
